relocate chunk files AGAIN
