+++ trilogy-vm/src/bytecode/chunk/builder.rs
@@ -1,42 +1,6 @@
+use super::Chunk;
 use crate::Procedure;
 use crate::{atom::AtomInterner, traits::Tags, Atom, Instruction, OpCode, Value};
-use std::collections::HashMap;
-use std::fmt::Debug;
-
-/// A chunk of independently compiled source code for this VM.
-#[derive(Clone)]
-pub struct Chunk {
-    labels: HashMap<String, u32>,
-    pub(crate) constants: Vec<Value>,
-    pub(crate) bytes: Vec<u8>,
-}
-
-impl Debug for Chunk {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        writeln!(f, "Constants:")?;
-        for (i, constant) in self.constants.iter().enumerate() {
-            writeln!(f, "{i}: {constant:?}")?;
-        }
-
-        let mut needs_break = false;
-        for (i, byte) in self.bytes.iter().enumerate() {
-            for (label, _) in self
-                .labels
-                .iter()
-                .filter(|(_, offset)| **offset == i as u32)
-            {
-                if needs_break {
-                    writeln!(f)?;
-                    needs_break = false;
-                }
-                writeln!(f, "{label}:")?;
-            }
-            write!(f, "{byte:02x}")?;
-            needs_break = true;
-        }
-        writeln!(f)
-    }
-}
 
 enum Parameter {
     Value(Value),
@@ -154,7 +118,7 @@ impl ChunkBuilder {
     /// Construct the [`Chunk`][] that was being built. Fails if any labels were referenced
     /// but not defined.
     pub fn build(mut self) -> Result<Chunk, ChunkError> {
-        let mut label_offsets = HashMap::new();
+        let mut label_offsets = std::collections::HashMap::new();
         let mut constants = vec![];
         let mut bytes = vec![];
